lib/pywpcare/requirements.txt
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
